🐛 [Bug] TicketSetup 유무 확인하는 로직 전부 False로 나가는 버그 발생 #1008 (#1009)

+++ gg-agenda-api/src/main/java/gg/agenda/api/user/ticket/controller/response/TicketCountResDto.java
@@ -6,10 +6,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 public class TicketCountResDto {
-	private int ticketCount;
+	private long ticketCount;
 	private boolean setupTicket;
 
-	public TicketCountResDto(int ticketCount, boolean setupTicket) {
+	public TicketCountResDto(long ticketCount, boolean setupTicket) {
 		this.ticketCount = ticketCount;
 		this.setupTicket = setupTicket;
 	}
